+++ packages/ui/src/settings-sidebar.ts
@@ -1,0 +1,3 @@
+/// <reference path="./shims-vue.d.ts" />
+
+export { default as LiliaSettingsSidebar } from "./layouts/SettingsSidebar.vue";