+++ packages/ui/src/shell-app.ts
@@ -1,0 +1,3 @@
+/// <reference path="./shims-vue.d.ts" />
+
+export { default as LiliaAppShell } from "./layouts/AppShell.vue";